Move useInviteEmailToSourcegraph
