--- client/web/src/auth/welcome/InviteCollaborators/InviteCollaborators.tsx
@@ -1,8 +1,0 @@
-import { Maybe } from '@sourcegraph/shared/src/graphql-operations'
-
-export interface InvitableCollaborator {
-    email: string
-    displayName: string
-    name: string
-    avatarURL: Maybe<string>
-}
